+++ GameEngineCore/src/objects/PlayerObject.cpp
@@ -33,7 +33,7 @@ namespace spacey{
 			regenHealth();
 
 			if (!shot.empty()){
-				for (int i = 0; i < shot.size(); i++){
+				for (unsigned int i = 0; i < shot.size(); i++){
 					shot[i].Fire();
 					if (shot[i].limit()){
 						shot.erase(shot.begin() + i);
@@ -69,7 +69,7 @@ namespace spacey{
 		}
 
 		void PlayerObject::checkRotation(Motion* motion){
-			direction = motion->angle;
+			direction = (int)motion->angle;
 
 			if (direction == 0){
 				direction = 1;
