+++ GameEngineCore/src/objects/Wall.h
@@ -10,7 +10,7 @@ namespace spacey{
 		class Wall : public BaseObject{
 		public:
 			Wall();
-			Wall(int width, int height, int x, int y, string type);			
+			Wall(int width, int height, float x, float y, string type);			
 		};
 	}
 }
